"Add configuration data and update class structure for plugins"
This commit adds configuration data for the "dreamy" addon and plugins in their corresponding yml files. The project's pom.xml files have been modified to include a new build resource for the plugins and addon, and new dependencies were added in the plugins' pom.xml file. Moreover, the Java main class for the plugins was updated from a simple HelloWorld program to a class that extends JavaPlugin, the standard base class for all plugins in Bukkit. Further, the .idea/misc.xml is updated to JDK_1_8 and a new .idea/vcs.xml file and .idea/discord.xml files are included for project setup. These changes aim to establish a structure for addon and plugin management and to align the development environment with the project requirements.